requirements.txt
alembic==1.11.1
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
black==23.7.0
cachetools==5.3.1
certifi==2023.7.22
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
distlib==0.3.7
exceptiongroup==1.1.2
filelock==3.12.2
flake8==3.9.2
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
identify==2.5.26
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.6.1
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pluggy==1.2.0
pprintpp==0.4.0
pre-commit==3.3.3
pycodestyle==2.7.0
pydocstyle==6.3.0
pyflakes==2.3.1
Pygments==2.15.1
PyJWT==2.8.0
pyproject-api==1.5.3
pytest==7.4.0
pytest-black==0.3.12
pytest-clarity==1.0.1
pytest-dotenv==0.5.2
pytest-flake8==1.1.0
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
six==1.16.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.0
toml==0.10.2
tomli==2.0.1
tox==4.6.4
typing_extensions==4.7.1
urllib3==2.0.4
virtualenv==20.24.1
Werkzeug==2.3.6
zipp==3.16.2
